#23 Add first pass at a page template
First pass at a marketing page template (since that’s all that’s in scope for this ticket), with some examples of how to use it.

Fixes for anything that broke along the way, like the Jest-based test that was from #32.

Better adaptation to current colors being used in the Figma design.